27 add jdbc repository implementation (#32)
* WIP: interval commit

* WIP: JPA interval

+++ embabel-database-agent/src/test/java/com/embabel/database/agent/service/CombinedServiceITest.java
@@ -65,11 +65,11 @@ public class CombinedServiceITest {
         //check again
         assertFalse(modelRepository.findAll().isEmpty());
         //get count
-        int baseline = modelRepository.count();
+        long baseline = modelRepository.count();
         //invoke bedrock
         bedrockModelParserService.loadModels();
         //get second count
-        int afterCount = modelRepository.count();
+        long afterCount = modelRepository.count();
         assertTrue(afterCount > baseline);
         //dump out
         try {
